+++ sdk/core/azure-core-experimental/src/main/java/com/azure/core/experimental/util/BinaryData.java
@@ -21,10 +21,10 @@ import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
 
 /**
- * This is binary representation of data from different sources. The example of such sources are {@link InputStream} ,
- * {@link Flux<ByteBuffer>} , {@link String} and byte array. One of the important API it provides is to serialize and
- * deserialize an {@link Object} into {@link BinaryData} given an {@link ObjectSerializer}.Following  are some examples
- * on how to use this class.
+ * This class is an abstraction over many different ways that binary data can be represented. For example
+ * {@link InputStream}, {@link Flux} of {@link ByteBuffer} , {@link String} and byte array. One of the important API it
+ * provides is to serialize and deserialize an {@link Object} into {@link BinaryData} given an {@link ObjectSerializer}.
+ * Following are some examples.
  *
  * <p><strong>Create an instance from Bytes</strong></p>
  * {@codesnippet com.azure.core.experimental.util.BinaryDocument.from#bytes}
@@ -37,12 +37,12 @@ import java.util.Arrays;
  *
  * @see ObjectSerializer
  */
-final public class BinaryData {
-    private byte[] data;
-    private static final ClientLogger LOGGER = new ClientLogger(BinaryData.class);
+public final class BinaryData {
+    private final byte[] data;
 
     BinaryData() {
         // This exists, so no one is able to create instance, user need to use static function to create instances.
+        this.data = null;
     }
 
     /**
@@ -71,7 +71,7 @@ final public class BinaryData {
      * @return {@link InputStream} representation of the data.
      */
     public Mono<InputStream> toStreamAsync() {
-        return Mono.fromCallable(() -> toStream()) ;
+        return Mono.fromCallable(() -> toStream());
     }
 
     /**
@@ -82,44 +82,46 @@ final public class BinaryData {
      * {@codesnippet com.azure.core.experimental.util.BinaryDocument.from#Stream}
      *
      * @param inputStream to read bytes from.
+     * @throws UncheckedIOException If any error in reading from {@link InputStream}.
      * @return {@link BinaryData} representing binary data.
      */
     public static BinaryData fromStream(InputStream inputStream) {
         final int bufferSize = 1024;
         try {
-            ByteArrayOutputStream buffer = new ByteArrayOutputStream();
+            ByteArrayOutputStream dataOutputBuffer = new ByteArrayOutputStream();
             int nRead;
             byte[] data = new byte[bufferSize];
             while ((nRead = inputStream.read(data, 0, data.length)) != -1) {
-                buffer.write(data, 0, nRead);
+                dataOutputBuffer.write(data, 0, nRead);
             }
 
-            return fromBytes(buffer.toByteArray());
+            return fromBytes(dataOutputBuffer.toByteArray());
         } catch (IOException ex) {
-            throw LOGGER.logExceptionAsError(new UncheckedIOException(ex));
+            ClientLogger logger = new ClientLogger(BinaryData.class);
+            throw logger.logExceptionAsError(new UncheckedIOException(ex));
         }
     }
 
     /**
-     * Asynchronously create {@link BinaryData} instance with given {@link InputStream} as source of data. The {@link InputStream} is
-     * not closed by this function.
+     * Asynchronously create {@link BinaryData} instance with given {@link InputStream} as source of data. The
+     * {@link InputStream} is not closed by this function.
      *
      * @param inputStream to read bytes from.
-     * @return {@link Mono<BinaryData>} representing binary data.
+     * @return {@link Mono} of {@link BinaryData} representing binary data.
      */
     public static Mono<BinaryData> fromStreamAsync(InputStream inputStream) {
-        return Mono.fromCallable(() -> fromStream(inputStream)) ;
+        return Mono.fromCallable(() -> fromStream(inputStream));
     }
 
     /**
-     * Create {@link BinaryData} instance with given {@link Flux<ByteBuffer>} as source of data. It will collect all
-     * the bytes from {@link ByteBuffer} into {@link BinaryData}.
+     * Create {@link BinaryData} instance with given {@link Flux} of {@link ByteBuffer} as source of data. It will
+     * collect all the bytes from {@link ByteBuffer} into {@link BinaryData}.
      *
      * <p><strong>Create an instance from String</strong></p>
      * {@codesnippet com.azure.core.experimental.util.BinaryDocument.from#Flux}
      *
      * @param data to use.
-     * @return {@link Mono<BinaryData>} representing binary data.
+     * @return {@link Mono} of {@link BinaryData} representing binary data.
      */
     public static Mono<BinaryData> fromFlux(Flux<ByteBuffer> data) {
         return FluxUtil.collectBytesInByteBufferStream(data)
@@ -134,6 +136,7 @@ final public class BinaryData {
      *
      * @param data to use.
      * @param charSet to use.
+     * @return {@link BinaryData} representing binary data.
      */
     public static BinaryData fromString(String data, Charset charSet) {
         return new BinaryData(data.getBytes(charSet));
@@ -144,6 +147,7 @@ final public class BinaryData {
      * {@link StandardCharsets#UTF_8} character set.
      *
      * @param data to use.
+     * @return {@link BinaryData} representing binary data.
      */
     public static BinaryData fromString(String data) {
         return new BinaryData(data.getBytes(StandardCharsets.UTF_8));
@@ -153,6 +157,7 @@ final public class BinaryData {
      * Create {@link BinaryData} instance with given byte array data.
      *
      * @param data to use.
+     * @return {@link BinaryData} representing binary data.
      */
     public static BinaryData fromBytes(byte[] data) {
         return new BinaryData(data);
@@ -163,6 +168,7 @@ final public class BinaryData {
      *
      * @param data The {@link Object} which needs to be serialized into bytes.
      * @param serializer to use for serializing the object.
+     * @return {@link BinaryData} representing binary data.
      */
     public static BinaryData fromObject(Object data, ObjectSerializer serializer) {
         ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
@@ -171,13 +177,15 @@ final public class BinaryData {
     }
 
     /**
-     * Serialize the given {@link Object} into {@link Mono<BinaryData>} using the provided {@link ObjectSerializer}.
+     * Serialize the given {@link Object} into {@link Mono} {@link BinaryData} using the provided
+     * {@link ObjectSerializer}.
      *
      * @param data The {@link Object} which needs to be serialized into bytes.
      * @param serializer to use for serializing the object.
+     * @return {@link Mono} of {@link BinaryData} representing binary data.
      */
     public static Mono<BinaryData> fromObjectAsync(Object data, ObjectSerializer serializer) {
-        return Mono.fromCallable(() -> fromObject(data, serializer)) ;
+        return Mono.fromCallable(() -> fromObject(data, serializer));
 
     }
 
@@ -216,6 +224,7 @@ final public class BinaryData {
      *
      * @param clazz representing the type of the Object.
      * @param serializer to use deserialize data into type.
+     * @param <T> Generic type that the data is deserialized into.
      * @return The {@link Object} of given type after deserializing the bytes.
      */
     public <T> T toObject(Class<T> clazz, ObjectSerializer serializer) {
@@ -233,9 +242,10 @@ final public class BinaryData {
      *
      * @param clazz representing the type of the Object.
      * @param serializer to use deserialize data into type.
+     * @param <T> Generic type that the data is deserialized into.
      * @return The {@link Object} of given type after deserializing the bytes.
      */
     public  <T> Mono<T> toObjectAsync(Class<T> clazz, ObjectSerializer serializer) {
-        return Mono.fromCallable(() -> toObject(clazz, serializer)) ;
+        return Mono.fromCallable(() -> toObject(clazz, serializer));
     }
 }